Popravljene neke stvari iz zadace 4 koje se sjecam da si spominjao
