+++ tv_shows_app/src/components/features/Containers/AllShowsContainer/AllShowsContainer.tsx
@@ -2,9 +2,10 @@
 import { LoadingScreen } from '@/components/shared/LoadingScreen/LoadingScreen';
 import { ShowsList } from '@/components/shared/shows/ShowsList';
 import { SidebarNavigation } from '@/components/shared/SidebarNavigation/SidebarNavigation';
-import { getMutator } from '@/fetchers/getMutator';
+import { getMutator } from '@/fetchers/mutators';
+import { swrKeys } from '@/fetchers/swrKeys';
 import { IShow } from '@/typings/show';
-import { Box, Flex, Spinner } from '@chakra-ui/react';
+import { Box, cookieStorageManager, Flex, Spinner } from '@chakra-ui/react';
 import { useParams } from 'next/navigation';
 import { title } from 'process';
 import useSWR from 'swr';
@@ -16,35 +17,24 @@ interface IGetAllShowsParams {
 }
 export const AllShowsContainer = () => {
   const { trigger } = useSWRMutation(
-    'https://tv-shows.infinum.academy/shows',
+    swrKeys.allShows,
     getMutator<IGetAllShowsParams>
   );
   async function getAllShows(params: IGetAllShowsParams) {
-    const response = await trigger(params);
-    return response.data;
+    return await trigger(params);
   }
 
-  const { data, error, isLoading } = useSWR(`/api/shows`, () =>
+  const { data, error, isLoading } = useSWR(`/api/shows`, async () =>
     getAllShows({ page: '2', items: '5' })
   );
   const showList = data?.shows || [];
   if (isLoading || !data) {
+    console.log(isLoading + ' ' + data);
     return <LoadingScreen />;
   }
 
   if (error) {
     return <div> Ajoj čini se da se nešto jaaaako loše desilo... </div>;
   }
-  return (
-    <main>
-      <Flex height="100vh">
-        <Box width="15%">
-          <SidebarNavigation />
-        </Box>
-        <Box width="85%">
-          <ShowsList shows={showList} />
-        </Box>
-      </Flex>
-    </main>
-  );
+  return <ShowsList shows={showList} />;
 };
