+++ tv_shows_app/src/components/features/Containers/TopShowsContainer/TopShowsContainer.tsx
@@ -2,8 +2,9 @@
 import { LoadingScreen } from '@/components/shared/LoadingScreen/LoadingScreen';
 import { ShowsList } from '@/components/shared/shows/ShowsList';
 import { SidebarNavigation } from '@/components/shared/SidebarNavigation/SidebarNavigation';
-import { getMutator } from '@/fetchers/getMutator';
+import { getMutator } from '@/fetchers/mutators';
 import { getTopShows } from '@/fetchers/shows';
+import { swrKeys } from '@/fetchers/swrKeys';
 import { Box, Flex, Spinner } from '@chakra-ui/react';
 import { useParams } from 'next/navigation';
 import useSWR from 'swr';
@@ -16,12 +17,11 @@ interface IGetAllShowsParams {
 
 export const TopShowsContainer = () => {
   const { trigger } = useSWRMutation(
-    'https://tv-shows.infinum.academy/shows/top_rated',
+    swrKeys.topRated,
     getMutator<IGetAllShowsParams>
   );
   async function getTopShows(params: IGetAllShowsParams) {
-    const response = await trigger(params);
-    return response.data;
+    return await trigger(params);
   }
 
   const { data, error, isLoading } = useSWR(`/api/top-rated`, () =>
@@ -35,16 +35,5 @@ export const TopShowsContainer = () => {
   if (error) {
     return <div> Ajoj čini se da se nešto jaaaako loše desilo... </div>;
   }
-  return (
-    <main>
-      <Flex height="100vh">
-        <Box width="15%">
-          <SidebarNavigation />
-        </Box>
-        <Box width="85%">
-          <ShowsList shows={showList} />
-        </Box>
-      </Flex>
-    </main>
-  );
+  return <ShowsList shows={showList} />;
 };
