+++ tv_shows_app/src/components/features/auth/LoginForm/LoginForm.tsx
@@ -1,10 +1,23 @@
 'use client';
-import { CustomInput } from '@/components/shared/auth/CustomInput';
+import { AuthRedirect } from '@/components/shared/auth/AuthRedirect';
 import { PasswordInput } from '@/components/shared/auth/PasswordInput';
 import { SuccessWindow } from '@/components/shared/auth/SuccessWindow';
-import { mutator } from '@/fetchers/mutator';
+import { LoadingScreen } from '@/components/shared/LoadingScreen/LoadingScreen';
+import { mutator } from '@/fetchers/mutators';
+import { swrKeys } from '@/fetchers/swrKeys';
 import { EmailIcon, LockIcon } from '@chakra-ui/icons';
-import { Button, Flex, Heading, Spinner, Text } from '@chakra-ui/react';
+import {
+  Button,
+  Flex,
+  FormControl,
+  Heading,
+  Input,
+  InputGroup,
+  InputLeftElement,
+  Spinner,
+  Text,
+} from '@chakra-ui/react';
+import { log } from 'console';
 import NextLink from 'next/link';
 import { useState } from 'react';
 import { useForm } from 'react-hook-form';
@@ -17,80 +30,97 @@ interface ILoginForm {
 
 export const LoginForm = () => {
   const [loggedIn, setLoggedIn] = useState(false);
+  const onAuth = (login: boolean) => {
+    if (login) setLoggedIn(true);
+    else setLoggedIn(false);
+  };
   const {
     register,
     handleSubmit,
     formState: { isSubmitting },
   } = useForm<ILoginForm>();
-  const { trigger } = useSWRMutation(
-    'https://tv-shows.infinum.academy/users/sign_in',
-    mutator<ILoginForm>
-  );
+  const { trigger } = useSWRMutation(swrKeys.login, mutator<ILoginForm>);
   const onLogin = async (data: ILoginForm) => {
     const response = await trigger(data);
     localStorage.setItem('tv-shows-uid', response.data.user.id);
-    localStorage.setItem('tv-shows-header', response.headers);
     setLoggedIn(true);
   };
+  console.log(loggedIn);
   return (
-    <Flex
-      direction="column"
-      alignItems="center"
-      justifyContent="center"
-      gap="4"
-      backgroundColor="pink.900"
-      height="100vh"
-    >
-      {loggedIn ? (
-        <SuccessWindow link="/shows" />
-      ) : (
-        <Flex
-          backgroundColor="pink.800"
-          direction="column"
-          alignItems="center"
-          as="form"
-          gap="4"
-          padding="10"
-          borderRadius="10"
-          onSubmit={handleSubmit(onLogin)}
-        >
-          <Heading color="white" size="md" marginBottom="8">
-            TV shows APP
-          </Heading>
-          <CustomInput
-            disable={isSubmitting}
-            thatPart={register('email')}
-            placeholder="Email"
-            type="email"
-            icon={<EmailIcon color="gray.300" />}
-            testId="email"
-          />
-          <PasswordInput
-            disable={isSubmitting}
-            thatPart={register('password')}
-            placeholder="Password"
-            icon={<LockIcon color="gray.300" />}
-            testId="password"
-          />
-          <Button
-            isLoading={isSubmitting}
-            loadingText="Logging in"
-            type="submit"
-            marginTop="8"
-            backgroundColor="orange.100"
-            _hover={{ bg: 'green.200' }}
-            borderRadius="20px"
+    <>
+      <AuthRedirect loginFunc={onAuth} />
+      <Flex
+        direction="column"
+        alignItems="center"
+        justifyContent="center"
+        gap="4"
+        backgroundColor="pink.900"
+        height="100vh"
+      >
+        {loggedIn ? (
+          <SuccessWindow link="/shows" message="Logged in!" />
+        ) : (
+          <Flex
+            backgroundColor="pink.800"
+            direction="column"
+            alignItems="center"
+            as="form"
+            gap="4"
+            padding="10"
+            borderRadius="10"
+            onSubmit={handleSubmit(onLogin)}
           >
-            Log in
-          </Button>
-          <Flex flexDirection="row" color="white" gap="1" fontSize="xs">
-            <Text> Don't have an account? </Text>
-            <Text as={NextLink} href="/register" textDecoration="underline">
-              Register
-            </Text>
+            <Heading color="white" size="md" marginBottom="8">
+              TV shows APP
+            </Heading>
+            <InputGroup>
+              <InputLeftElement
+                pointerEvents="none"
+                children={<EmailIcon color="gray.300" />}
+                fontSize="md"
+              />
+              <FormControl textColor="white">
+                <Input
+                  isDisabled={isSubmitting}
+                  {...register('email')}
+                  paddingLeft="35px"
+                  placeholder="Email"
+                  type="email"
+                  color="white"
+                  _placeholder={{ color: 'white' }}
+                  borderRadius="20px"
+                  size="md"
+                  data-testid="email"
+                />
+              </FormControl>
+            </InputGroup>
+            <PasswordInput
+              disable={isSubmitting}
+              thatPart={register('password')}
+              placeholder="Password"
+              icon={<LockIcon color="gray.300" />}
+              testId="password"
+            />
+            <Button
+              isLoading={isSubmitting}
+              loadingText="Logging in"
+              type="submit"
+              marginTop="8"
+              backgroundColor="orange.100"
+              _hover={{ bg: 'green.200' }}
+              borderRadius="20px"
+            >
+              Log in
+            </Button>
+            <Flex flexDirection="row" color="white" gap="1" fontSize="xs">
+              <Text> Don't have an account? </Text>
+              <Text as={NextLink} href="/register" textDecoration="underline">
+                Register
+              </Text>
+            </Flex>
           </Flex>
-        </Flex>
-      )}
-    </Flex>
+        )}
+      </Flex>
+    </>
   );
 };
