+++ tv_shows_app/src/components/features/auth/RegisterForm/RegisterForm.tsx
@@ -1,14 +1,25 @@
 'use client';
-import { CustomInput } from '@/components/shared/auth/CustomInput';
-import { mutator } from '@/fetchers/mutator';
+import { mutator } from '@/fetchers/mutators';
 import { EmailIcon, LockIcon, RepeatIcon } from '@chakra-ui/icons';
 import NextLink from 'next/link';
-import { Box, Button, Flex, Heading, Spinner, Text } from '@chakra-ui/react';
+import {
+  Box,
+  Button,
+  Flex,
+  FormControl,
+  Heading,
+  Input,
+  InputGroup,
+  InputLeftElement,
+  Spinner,
+  Text,
+} from '@chakra-ui/react';
 import { useForm } from 'react-hook-form';
 import useSWRMutation from 'swr/mutation';
 import { useState } from 'react';
 import { SuccessWindow } from '@/components/shared/auth/SuccessWindow';
 import { PasswordInput } from '@/components/shared/auth/PasswordInput';
+import { swrKeys } from '@/fetchers/swrKeys';
 
 interface IRegisterForm {
   email: string;
@@ -24,15 +35,11 @@ export const RegisterForm = () => {
     formState: { isSubmitting },
   } = useForm<IRegisterForm>();
   const [errorMesssage, setErrorMessage] = useState('');
-  const { trigger } = useSWRMutation(
-    'https://tv-shows.infinum.academy/users',
-    mutator<IRegisterForm>,
-    {
-      onSuccess: () => {
-        setRegistred(true);
-      },
-    }
-  );
+  const { trigger } = useSWRMutation(swrKeys.register, mutator<IRegisterForm>, {
+    onSuccess: () => {
+      setRegistred(true);
+    },
+  });
   const onRegister = async (data: IRegisterForm) => {
     if (data.password.length < 8) {
       setErrorMessage('Password must be at least 8 characters');
@@ -55,7 +62,7 @@ export const RegisterForm = () => {
       height="100vh"
     >
       {registred ? (
-        <SuccessWindow link="/login" />
+        <SuccessWindow link="/login" message="Successfully registred!" />
       ) : (
         <Flex
           backgroundColor="pink.800"
@@ -70,14 +77,27 @@ export const RegisterForm = () => {
           <Heading color="white" size="md" marginBottom="8">
             TV shows APP
           </Heading>
-          <CustomInput
-            disable={isSubmitting}
-            thatPart={register('email')}
-            placeholder="Email"
-            type="email"
-            icon={<EmailIcon color="gray.300" />}
-            testId="email"
-          />
+          <InputGroup>
+            <InputLeftElement
+              pointerEvents="none"
+              children={<EmailIcon color="gray.300" />}
+              fontSize="md"
+            />
+            <FormControl textColor="white">
+              <Input
+                isDisabled={isSubmitting}
+                {...register('email')}
+                paddingLeft="35px"
+                placeholder="Email"
+                type="email"
+                color="white"
+                _placeholder={{ color: 'white' }}
+                borderRadius="20px"
+                size="md"
+                data-testid="email"
+              />
+            </FormControl>
+          </InputGroup>
           <PasswordInput
             disable={isSubmitting}
             thatPart={register('password')}
