+++ tv_shows_app/src/components/shared/auth/SuccessWindow.tsx
@@ -3,9 +3,10 @@ import NextLink from 'next/link';
 
 interface ISuccessWindow {
   link: string;
+  message: string;
 }
 
-export const SuccessWindow = ({ link }: ISuccessWindow) => (
+export const SuccessWindow = ({ link, message }: ISuccessWindow) => (
   <Flex
     backgroundColor="pink.800"
     color="white"
@@ -17,7 +18,7 @@ export const SuccessWindow = ({ link }: ISuccessWindow) => (
     justifyContent="center"
     gap="5"
   >
-    <Text>Success!</Text>
+    <Text>{message}</Text>
     <Button
       as={NextLink}
       href={link}
